feat: Create HTML report
remove mailbox

Issue #7
