+++ pmlib/console.py
@@ -17,21 +17,24 @@
 #
 
 from optparse import OptionParser
-from typing import Union
 
 import pmlib
 import pmlib.log
 
 from pmlib import config
-from pmlib.mailbox import Mailbox
+
+from pmlib.convert import Convert
+from pmlib.hierachy import Hierarchy
+from pmlib.item import Item, sort_items
+from pmlib.types import Entry
+
 from pmlib.utils import create_folder
 
 
 class Console(object):
 
     def __init__(self):
-
-        self.mailbox: Union[Mailbox, None] = None
+        self._converter: Convert = Convert()
 
         usage = "usage: %prog [options] arg1 arg2"
         self.parser: OptionParser = OptionParser(usage=usage)
@@ -46,6 +49,76 @@ class Console(object):
         self.parser.add_option("-t", "--target", help="target path for export", type="string", metavar="<FOLDER>",
                                default="")
         return
+
+    def _create_folder(self, item: Item) -> bool:
+        item.set_target()
+
+        if item.type is Entry.folder:
+            return True
+
+        check = create_folder(item.target)
+        if check is False:
+            pmlib.log.error("Unable to create target folder: {0:s}".format(item.target))
+            return False
+
+        for _item in item.children:
+            check = self._create_folder(_item)
+            if check is False:
+                return False
+
+        return True
+
+    def _convert_item(self, item: Item) -> bool:
+        attr = self._converter.get_converter(item.data.type)
+        if attr is None:
+            text = "Unable to convert folder {0:s} with type {1:s} to {2:s}".format(item.name,
+                                                                                    item.data.type.name,
+                                                                                    pmlib.config.target_type.name)
+            pmlib.log.warn("Mailbox", text)
+            return True
+
+        converter = attr()
+        check = converter.prepare(item)
+        if check is False:
+            return False
+
+        check = converter.run()
+        if check is False:
+            return False
+
+        check = converter.close()
+        if check is False:
+            return False
+
+        for _error in item.report.error:
+            pmlib.log.warn(item.name, _error.text)
+
+        return True
+
+    def _export_item(self, item: Item) -> bool:
+
+        if item.type is Entry.folder:
+            check = self._convert_item(item)
+            if check is False:
+                return False
+        else:
+            pmlib.log.inform("TRAY", item.full_name)
+
+            # first convert folder
+            for _item in sorted(item.children, key=sort_items):
+                if _item.type is Entry.folder:
+                    check = self._export_item(_item)
+                    if check is False:
+                        return False
+
+            # then trays
+            for _item in sorted(item.children, key=sort_items):
+                if _item.type is not Entry.folder:
+                    check = self._export_item(_item)
+                    if check is False:
+                        return False
+
+        return True
 
     def prepare(self) -> bool:
         (options, args) = self.parser.parse_args()
@@ -65,22 +138,33 @@ class Console(object):
             pmlib.log.error("Unable to create target folder!")
             return False
 
-        self.mailbox = Mailbox()
         return True
 
     def run(self) -> bool:
-        check = self.mailbox.init()
+        self._converter.init()
+
+        hierarchy = Hierarchy()
+
+        count = hierarchy.parse()
+        if count == 0:
+            return False
+
+        hierarchy.sort()
+
+        root = pmlib.data.root
+
+        check = self._create_folder(pmlib.data.root)
         if check is False:
             return False
 
-        check = self.mailbox.convert()
+        check = self._export_item(pmlib.data.root)
         if check is False:
             return False
 
         return True
 
-    def close(self) -> bool:
+    @staticmethod
+    def close() -> bool:
 
         check = True
-        # check = self.mailbox.report()
         return check
